requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.2.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
django-cors-headers==4.3.1
python-dateutil==2.8.2
newsapi-python==0.2.7
feedparser==6.0.10
channels==4.0.0
channels-redis==4.1.0
pydantic==2.5.0

